pyproject.toml
[tool.poetry]
package-mode = false
name = "amazon-bedrock-agent-chat-completions-proxy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "3.0.2"
boto3 = "1.34.69"
python-dotenv = "1.0.1"
requests = "2.31.0"
websockets = "^14.2"
sounddevice = "^0.5.1"
numpy = "^2.2.2"
scipy = "^1.15.1"
openai = "^1.77.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
